Re-organize the source code tree structure.
Add these sub-directories:
  - enemies/
  - bullets/
  - menus/
  - maps/
  - shared/
