+++ source/bullets/bullet4_mines.ts
