+++ source/music.ts
@@ -1,4 +1,4 @@
-import { getMusicVolume } from "./options.js";
+import { getMusicVolume } from "./shared/options.js";
 
 const SONG_NAMES = ["scumm_bar", "space_ship_1"];
 let CURRENT_SONG = 0;
