+++ source/shared/constants.ts
@@ -1,7 +1,7 @@
-import EnemyMoveHorizontally from "../enemy_move_horizontally.js";
-import EnemyKamikaze from "../enemy_kamikaze.js";
-import EnemyRotateAround from "../enemy_rotate_around.js";
-import EnemyRocks from "../enemy_rocks.js";
+import EnemyMoveHorizontally from "../enemies/enemy_move_horizontally.js";
+import EnemyKamikaze from "../enemies/enemy_kamikaze.js";
+import EnemyRotateAround from "../enemies/enemy_rotate_around.js";
+import EnemyRocks from "../enemies/enemy_rocks.js";
 
 export const EnemyMapping = {
     EnemyMoveHorizontally: EnemyMoveHorizontally,
