+++ source/shared/message.ts
