+++ source/shared/utilities.ts
@@ -1,4 +1,4 @@
-import { CANVAS, b2Vec2 } from "./main.js";
+import { CANVAS, b2Vec2 } from "../main.js";
 
 export const EVENT_KEY = {
     backspace: 8,
@@ -96,29 +96,6 @@ export function centerElement(element, refElement?) {
         top: top + "px",
         left: left + "px",
     });
-}
-
-/*
- * Used for 'class' inheritance (search prototypal inheritance)
- */
-function OBJECT(o) {
-    function F() {}
-
-    F.prototype = o;
-
-    return new F();
-}
-
-/*
- * Used for 'class' inheritance (search for parasitic combination inheritance)
- */
-
-function INHERIT_PROTOTYPE(derivedClass, baseClass) {
-    var prototype = OBJECT(baseClass.prototype);
-
-    prototype.constructor = derivedClass;
-
-    derivedClass.prototype = prototype;
 }
 
 /*
